fix: Resolve duplicate modal issue on vote reveal
- Move VotingResultsModal state management to RoomPage level
- Share single modal instance between desktop and mobile layouts
- Remove duplicate modal rendering in ResponsivePlayerLayout
- Add modal control props to ResponsivePlayerLayout interface
- Implement proper modal opening coordination between layouts

Fixes the issue where two identical VotingResultsModal dialogs
would open simultaneously when revealing votes due to RoomPage
rendering ResponsivePlayerLayout twice (desktop and mobile).
